.xinitrc: Rotate and move DisplayPort-1 and 4
For some reason the AMD card in presci (which has 4 mini-DP outputs)
sometimes comes up as outputs 0, 1, 2, 3 and other times comes up as
outputs 0, 4, 5, 6.  When it comes up as 0, 4, 5, 6 then there are no
outputs 1, 2, 3 on the system, so it's unclear why this happens.  But
since we want the 2nd monitor to always rotate and be to the right just
do the same things to outputs 1 and 4 since we only have 2 monitors.